Добавил три новых окна на Swing дизайнере. Начал разработку нового Game клиента.
