--- src/main/java/MKRefactor/Character/Baraka.java
@@ -1,24 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package MKRefactor.Character;
-
-import MKRefactor.Game.Player;
-
-/**
- *
- * @author Мария
- */
-public class Baraka extends Player {
-    
-    public Baraka(int level, int health, int  damage, int attack){
-        super (level, health, damage, attack);
-    }
-    
-    @Override
-    public String getName(){
-        return "Baraka";
-    }
-    
-}
