--- src/main/java/MKRefactor/Character/SubZero.java
@@ -1,23 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package MKRefactor.Character;
-
-import MKRefactor.Game.Player;
-
-/**
- *
- * @author Мария
- */
-public class SubZero extends Player {
-    
-    public SubZero(int level, int health, int damage , int attack){
-        super (level, health, damage, attack);
-    }
-    
-    @Override
-    public String getName(){
-        return "Sub-Zero";
-    }
-}
