--- src/main/java/MKRefactor/Fabric/EnemyFabricInterface.java
@@ -1,15 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
- */
-package MKRefactor.Fabric;
-
-import MKRefactor.Game.Player;
-
-/**
- *
- * @author Мария
- */
-public interface EnemyFabricInterface {
-    public Player create();
-}
