--- src/main/java/MKRefactor/MortalKombatBVersion.java
@@ -1,25 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Main.java to edit this template
- */
-package MKRefactor;
-
-import MKRefactor.GUI.JFrames;
-
-/**
- *
- * @author Мария
- */
-public class MortalKombatBVersion {
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String[] args) {
-        
-        JFrames app = new JFrames();
-        app.setLocationRelativeTo(null);
-        app.setVisible(true);
-    }
-    
-}
