+++ src/main/java/MortalCombat/Character/SonyaBlade.java
@@ -1,0 +1,23 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package MortalCombat.Character;
+
+import MortalCombat.Game.Player;
+
+/**
+ *
+ * @author Мария
+ */
+public class SonyaBlade extends Player {
+    
+    public SonyaBlade (int level, int health, int  damage, int attack){
+        super (level, health, damage, attack);
+    }
+    
+    @Override
+    public String getName(){
+        return "Sonya Blade";
+    }
+}
