+++ src/main/java/MortalCombat/Fabric/EnemyFabricInterface.java
@@ -1,0 +1,15 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
+ */
+package MortalCombat.Fabric;
+
+import MortalCombat.Game.Player;
+
+/**
+ *
+ * @author Мария
+ */
+public interface EnemyFabricInterface {
+    public Player create();
+}
