+++ src/main/java/MortalCombat/Fabric/SubZeroFabric.java
@@ -1,0 +1,24 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package MortalCombat.Fabric;
+
+
+import MortalCombat.Game.Player;
+import MortalCombat.Character.SubZero;
+
+/**
+ *
+ * @author Мария
+ */
+public class SubZeroFabric implements EnemyFabricInterface {
+
+    @Override
+    public Player create() {
+        Player enemy;
+        enemy = new SubZero(1, 60, 16, 1);
+        return enemy;
+    }
+
+}
