+++ src/main/java/MortalCombat/GUI/JFrames.java
@@ -1,0 +1,1256 @@
+package MortalCombat.GUI;
+
+import MortalCombat.Game.*;
+
+import java.awt.*;
+import java.awt.image.BufferedImage;
+import java.io.IOException;
+import java.net.URL;
+import java.util.Objects;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+import javax.imageio.ImageIO;
+import javax.swing.*;
+
+import static MortalCombat.Handlers.ImageResizer.resizeImage;
+
+/**
+ * Класс пользовательского интерфейса
+ * @see javax.swing.JFrame
+ * @author Мария
+ */
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
+ */
+
+
+/**
+ * Класс пользовательского интерфейса
+ * @see javax.swing.JFrame
+ * @author Мария
+ */
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+
+public class JFrames extends javax.swing.JFrame {
+    OldGame oldGame = new OldGame();
+    Human human = null;
+    Player enemy = null;
+    Items[] items = new Items[3];
+    /**
+     * Creates new form JFrame
+     */
+    public JFrames() {
+        initComponents();
+        try {
+            oldGame.readFromExcel();
+        } catch (IOException ex) {
+            Logger.getLogger(JFrames.class.getName()).log(Level.SEVERE, null, ex);
+        }
+        oldGame.writeToTable(jTable_records);
+        buttonGroup_items.add(jRadioButton_small_potion);
+        buttonGroup_items.add(jRadioButton_big_potion);
+        buttonGroup_items.add(jRadioButton_resurrection_cross);
+        items[0]=new Items("Малое зелье лечения",0);
+        items[1]=new Items("Большое зелье лечения",0);
+        items[2]=new Items("Крест возрождения",0);
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        jFrame_fight = new javax.swing.JFrame();
+        jPanel_fight = new javax.swing.JPanel();
+        jLabel_self_image = new javax.swing.JLabel();
+        jButton_attack = new javax.swing.JButton();
+        jButton_defend = new javax.swing.JButton();
+        jProgressBar_self = new javax.swing.JProgressBar();
+        jLabel_enemy_image = new javax.swing.JLabel();
+        jProgressBar_enemy = new javax.swing.JProgressBar();
+        jLabel_enemy_name = new javax.swing.JLabel();
+        jLabel_enemy_damage = new javax.swing.JLabel();
+        jLabel_self_damage = new javax.swing.JLabel();
+        jLabel_self_damage_value = new javax.swing.JLabel();
+        jLabel_enemy_damage_value = new javax.swing.JLabel();
+        jLabel_self_level = new javax.swing.JLabel();
+        jLabel_fight = new javax.swing.JLabel();
+        jLabel_self_health = new javax.swing.JLabel();
+        jLabel_enemy_health = new javax.swing.JLabel();
+        jLabel_experience = new javax.swing.JLabel();
+        jLabel_points = new javax.swing.JLabel();
+        jLabel_experience_value = new javax.swing.JLabel();
+        jLabel_points_value = new javax.swing.JLabel();
+        jLabel_enemy_level = new javax.swing.JLabel();
+        jLabel_turn = new javax.swing.JLabel();
+        jLabel_what_happened = new javax.swing.JLabel();
+        jLabel_Kitana = new javax.swing.JLabel();
+        jLabel29 = new javax.swing.JLabel();
+        jButton_items = new javax.swing.JButton();
+        jLabel_location = new javax.swing.JLabel();
+        jDialog_end_of_fight = new javax.swing.JDialog();
+        jPanel_end_of_fight = new javax.swing.JPanel();
+        jLabel_end_of_fight = new javax.swing.JLabel();
+        jButton_continue = new javax.swing.JButton();
+        jDialog_win_in_top = new javax.swing.JDialog();
+        jPanel_win_in_top = new javax.swing.JPanel();
+        jLabel_win_message_in_top = new javax.swing.JLabel();
+        jTextField_win_in_top = new javax.swing.JTextField();
+        jLabel_in_top_message1 = new javax.swing.JLabel();
+        jLabel_in_top_message2 = new javax.swing.JLabel();
+        jButton_finish_game_in_top = new javax.swing.JButton();
+        jDialog_records_table = new javax.swing.JDialog();
+        jPanel_records_table = new javax.swing.JPanel();
+        jLabel_records_table = new javax.swing.JLabel();
+        jScrollPane_records_table = new javax.swing.JScrollPane();
+        jTable_records = new javax.swing.JTable();
+        jButton_close_records_table = new javax.swing.JButton();
+        jDialog_win_outside_top = new javax.swing.JDialog();
+        jPanel_win_outside_top = new javax.swing.JPanel();
+        jLabel_win_message_outside_top = new javax.swing.JLabel();
+        jLabel_outside_top_message = new javax.swing.JLabel();
+        jButton_finish_game_outside_top = new javax.swing.JButton();
+        jDialog_bag_with_items = new javax.swing.JDialog();
+        jPanel_bag_with_items = new javax.swing.JPanel();
+        jLabel_bag_of_items = new javax.swing.JLabel();
+        jRadioButton_small_potion = new javax.swing.JRadioButton();
+        jRadioButton_big_potion = new javax.swing.JRadioButton();
+        jRadioButton_resurrection_cross = new javax.swing.JRadioButton();
+        jButton_use_item = new javax.swing.JButton();
+        buttonGroup_items = new javax.swing.ButtonGroup();
+        jDialog_unavailable_item = new javax.swing.JDialog();
+        jPanel_unavailable_item = new javax.swing.JPanel();
+        jLabel_unavailable_item1 = new javax.swing.JLabel();
+        jLabel_unavailable_item2 = new javax.swing.JLabel();
+        jButton_unavailable_item = new javax.swing.JButton();
+        jDialog_locations = new javax.swing.JDialog();
+        jPanel_locations = new javax.swing.JPanel();
+        jLabel_locations = new javax.swing.JLabel();
+        jTextField_locations = new javax.swing.JTextField();
+        jButton_locations_start = new javax.swing.JButton();
+        jPanel_start = new javax.swing.JPanel();
+        jLabel_mortal_kombat = new javax.swing.JLabel();
+        jButton_start_game = new javax.swing.JButton();
+        jButton_watch_results = new javax.swing.JButton();
+        jLabel3 = new javax.swing.JLabel();
+
+        jPanel_fight.setBackground(new java.awt.Color(255, 255, 255));
+
+        jButton_attack.setBackground(new java.awt.Color(255, 0, 0));
+        jButton_attack.setFont(new java.awt.Font("Comic Sans MS", 1, 12)); // NOI18N
+        jButton_attack.setText("Атаковать");
+        jButton_attack.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton_attackActionPerformed(evt);
+            }
+        });
+
+        jButton_defend.setBackground(new java.awt.Color(255, 204, 0));
+        jButton_defend.setFont(new java.awt.Font("Comic Sans MS", 1, 12)); // NOI18N
+        jButton_defend.setText("Защититься");
+        jButton_defend.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton_defendActionPerformed(evt);
+            }
+        });
+
+        jProgressBar_self.setBackground(new java.awt.Color(204, 204, 204));
+        jProgressBar_self.setForeground(new java.awt.Color(51, 255, 51));
+        jProgressBar_self.setMaximum(80);
+        jProgressBar_self.setMinimum(-1);
+
+        jProgressBar_enemy.setBackground(new java.awt.Color(204, 204, 204));
+        jProgressBar_enemy.setForeground(new java.awt.Color(0, 255, 0));
+        jProgressBar_enemy.setMinimum(-1);
+
+        jLabel_enemy_name.setBackground(new java.awt.Color(0, 0, 0));
+        jLabel_enemy_name.setFont(new java.awt.Font("Comic Sans MS", 2, 14)); // NOI18N
+
+        jLabel_enemy_damage.setBackground(new java.awt.Color(255, 255, 255));
+        jLabel_enemy_damage.setFont(new java.awt.Font("Comic Sans MS", 1, 14)); // NOI18N
+        jLabel_enemy_damage.setText("Damage");
+
+        jLabel_self_damage.setBackground(new java.awt.Color(255, 255, 255));
+        jLabel_self_damage.setFont(new java.awt.Font("Comic Sans MS", 1, 14)); // NOI18N
+        jLabel_self_damage.setText("Damage");
+
+        jLabel_self_damage_value.setBackground(new java.awt.Color(255, 255, 255));
+        jLabel_self_damage_value.setFont(new java.awt.Font("Comic Sans MS", 1, 14)); // NOI18N
+        jLabel_self_damage_value.setForeground(new java.awt.Color(255, 0, 0));
+        jLabel_self_damage_value.setText("16");
+
+        jLabel_enemy_damage_value.setBackground(new java.awt.Color(255, 255, 255));
+        jLabel_enemy_damage_value.setFont(new java.awt.Font("Comic Sans MS", 1, 14)); // NOI18N
+        jLabel_enemy_damage_value.setForeground(new java.awt.Color(255, 0, 0));
+        jLabel_enemy_damage_value.setText("16");
+
+        jLabel_self_level.setBackground(new java.awt.Color(255, 255, 255));
+        jLabel_self_level.setFont(new java.awt.Font("Comic Sans MS", 0, 12)); // NOI18N
+        jLabel_self_level.setText("0 уровень");
+
+        jLabel_fight.setBackground(new java.awt.Color(255, 255, 255));
+        jLabel_fight.setFont(new java.awt.Font("Comic Sans MS", 3, 36)); // NOI18N
+        jLabel_fight.setForeground(new java.awt.Color(0, 0, 153));
+        jLabel_fight.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        jLabel_fight.setText("FIGHT");
+        jLabel_fight.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
+
+        jLabel_self_health.setBackground(new java.awt.Color(255, 255, 255));
+        jLabel_self_health.setFont(new java.awt.Font("Comic Sans MS", 0, 14)); // NOI18N
+        jLabel_self_health.setForeground(new java.awt.Color(102, 102, 102));
+        jLabel_self_health.setText("80/80");
+
+        jLabel_enemy_health.setBackground(new java.awt.Color(255, 255, 255));
+        jLabel_enemy_health.setFont(new java.awt.Font("Comic Sans MS", 0, 14)); // NOI18N
+        jLabel_enemy_health.setForeground(new java.awt.Color(102, 102, 102));
+        jLabel_enemy_health.setText("80/80");
+
+        jLabel_experience.setBackground(new java.awt.Color(255, 255, 255));
+        jLabel_experience.setFont(new java.awt.Font("Comic Sans MS", 1, 18)); // NOI18N
+        jLabel_experience.setForeground(new java.awt.Color(128, 92, 31));
+        jLabel_experience.setText("experience");
+
+        jLabel_points.setBackground(new java.awt.Color(255, 255, 255));
+        jLabel_points.setFont(new java.awt.Font("Comic Sans MS", 1, 18)); // NOI18N
+        jLabel_points.setForeground(new java.awt.Color(128, 92, 31));
+        jLabel_points.setText("points");
+
+        jLabel_experience_value.setBackground(new java.awt.Color(255, 255, 255));
+        jLabel_experience_value.setFont(new java.awt.Font("Comic Sans MS", 0, 14)); // NOI18N
+        jLabel_experience_value.setForeground(new java.awt.Color(113, 90, 16));
+        jLabel_experience_value.setText("0/40");
+
+        jLabel_points_value.setBackground(new java.awt.Color(255, 255, 255));
+        jLabel_points_value.setFont(new java.awt.Font("Comic Sans MS", 0, 14)); // NOI18N
+        jLabel_points_value.setForeground(new java.awt.Color(113, 90, 16));
+        jLabel_points_value.setText("00");
+
+        jLabel_enemy_level.setBackground(new java.awt.Color(255, 255, 255));
+        jLabel_enemy_level.setFont(new java.awt.Font("Comic Sans MS", 0, 12)); // NOI18N
+        jLabel_enemy_level.setText("1 уровень");
+
+        jLabel_turn.setFont(new java.awt.Font("Comic Sans MS", 0, 16)); // NOI18N
+        jLabel_turn.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+
+        jLabel_what_happened.setFont(new java.awt.Font("Comic Sans MS", 1, 12)); // NOI18N
+        jLabel_what_happened.setForeground(new java.awt.Color(204, 0, 0));
+        jLabel_what_happened.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+
+        jLabel_Kitana.setFont(new java.awt.Font("Comic Sans MS", 2, 14)); // NOI18N
+        jLabel_Kitana.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        jLabel_Kitana.setText("Ваня-Убийца");
+
+        jLabel29.setFont(new java.awt.Font("Comic Sans MS", 0, 16)); // NOI18N
+        jLabel29.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+
+        jButton_items.setBackground(new java.awt.Color(174, 183, 106));
+        jButton_items.setFont(new java.awt.Font("Comic Sans MS", 1, 12)); // NOI18N
+        jButton_items.setText("Предметы");
+        jButton_items.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton_itemsActionPerformed(evt);
+            }
+        });
+
+        jLabel_location.setForeground(new java.awt.Color(255, 102, 0));
+
+        javax.swing.GroupLayout jPanel_fightLayout = new javax.swing.GroupLayout(jPanel_fight);
+        jPanel_fight.setLayout(jPanel_fightLayout);
+        jPanel_fightLayout.setHorizontalGroup(
+                jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                .addGap(20, 20, 20)
+                                .addGroup(jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                                        .addComponent(jLabel_enemy_health)
+                                        .addComponent(jLabel_enemy_level))
+                                .addGap(18, 18, 18)
+                                .addGroup(jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                .addGroup(jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                                .addComponent(jLabel_enemy_damage, javax.swing.GroupLayout.PREFERRED_SIZE, 62, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                                                .addComponent(jLabel_enemy_damage_value, javax.swing.GroupLayout.PREFERRED_SIZE, 40, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                                .addGap(10, 10, 10)
+                                                                .addComponent(jLabel_enemy_name, javax.swing.GroupLayout.PREFERRED_SIZE, 162, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                                        .addGroup(jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                                                                .addComponent(jLabel_fight, javax.swing.GroupLayout.PREFERRED_SIZE, 205, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                                .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                                        .addComponent(jProgressBar_enemy, javax.swing.GroupLayout.PREFERRED_SIZE, 162, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                                        .addGroup(jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                                                                .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                                                        .addGap(92, 92, 92)
+                                                                                        .addComponent(jLabel_points)
+                                                                                        .addGap(61, 61, 61)
+                                                                                        .addComponent(jLabel_experience))
+                                                                                .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                                                        .addGap(111, 111, 111)
+                                                                                        .addComponent(jLabel_points_value)
+                                                                                        .addGap(108, 108, 108)
+                                                                                        .addComponent(jLabel_experience_value))))))
+                                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 38, Short.MAX_VALUE))
+                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                .addComponent(jLabel_enemy_image, javax.swing.GroupLayout.PREFERRED_SIZE, 189, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                .addGroup(jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                                .addGap(69, 69, 69)
+                                                                .addComponent(jLabel_turn, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                                .addGap(0, 0, Short.MAX_VALUE))
+                                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                                .addGap(26, 26, 26)
+                                                                .addGroup(jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                                                                        .addComponent(jLabel_what_happened, javax.swing.GroupLayout.PREFERRED_SIZE, 237, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                                        .addComponent(jLabel29, javax.swing.GroupLayout.PREFERRED_SIZE, 227, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 46, Short.MAX_VALUE))))
+                                        .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel_fightLayout.createSequentialGroup()
+                                                .addGap(0, 0, Short.MAX_VALUE)
+                                                .addComponent(jButton_items, javax.swing.GroupLayout.PREFERRED_SIZE, 110, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                .addGap(46, 46, 46)))
+                                .addGroup(jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                .addComponent(jLabel_self_level)
+                                                .addGap(26, 26, 26))
+                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                .addGroup(jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                                .addComponent(jProgressBar_self, javax.swing.GroupLayout.PREFERRED_SIZE, 177, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                                .addGap(18, 18, 18)
+                                                                .addComponent(jLabel_self_health))
+                                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                                .addComponent(jLabel_self_damage, javax.swing.GroupLayout.PREFERRED_SIZE, 62, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                                                .addComponent(jLabel_self_damage_value, javax.swing.GroupLayout.PREFERRED_SIZE, 40, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                                                .addGap(45, 45, 45))
+                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                .addComponent(jLabel_self_image, javax.swing.GroupLayout.PREFERRED_SIZE, 245, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                .addGap(60, 60, 60))
+                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                .addComponent(jLabel_Kitana, javax.swing.GroupLayout.PREFERRED_SIZE, 118, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                .addGap(144, 144, 144))
+                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                .addComponent(jButton_attack, javax.swing.GroupLayout.PREFERRED_SIZE, 110, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                .addGap(46, 46, 46)
+                                                .addComponent(jButton_defend, javax.swing.GroupLayout.PREFERRED_SIZE, 109, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                .addGap(68, 68, 68))))
+                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                .addContainerGap()
+                                .addComponent(jLabel_location, javax.swing.GroupLayout.PREFERRED_SIZE, 173, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+        );
+        jPanel_fightLayout.setVerticalGroup(
+                jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel_fightLayout.createSequentialGroup()
+                                .addGroup(jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                .addContainerGap()
+                                                .addComponent(jLabel_location, javax.swing.GroupLayout.PREFERRED_SIZE, 46, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                                .addGroup(jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                                                        .addComponent(jProgressBar_self, javax.swing.GroupLayout.DEFAULT_SIZE, 32, Short.MAX_VALUE)
+                                                        .addComponent(jProgressBar_enemy, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                                .addGap(8, 8, 8)
+                                                                .addComponent(jLabel_self_health))
+                                                        .addComponent(jLabel_enemy_health, javax.swing.GroupLayout.Alignment.TRAILING)))
+                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                .addGap(23, 23, 23)
+                                                .addComponent(jLabel_fight, javax.swing.GroupLayout.PREFERRED_SIZE, 64, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                .addGap(38, 38, 38)
+                                                .addGroup(jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                                                        .addComponent(jLabel_experience)
+                                                        .addComponent(jLabel_points))
+                                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 8, Short.MAX_VALUE)
+                                                .addGroup(jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                                                        .addComponent(jLabel_experience_value)
+                                                        .addComponent(jLabel_points_value))))
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                .addGroup(jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                        .addComponent(jLabel_self_damage)
+                                        .addGroup(jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                                                .addComponent(jLabel_enemy_damage)
+                                                .addComponent(jLabel_enemy_damage_value)
+                                                .addComponent(jLabel_self_damage_value)))
+                                .addGap(27, 27, 27)
+                                .addGroup(jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                .addGroup(jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                                        .addComponent(jLabel_enemy_image, javax.swing.GroupLayout.PREFERRED_SIZE, 275, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                                .addGap(25, 25, 25)
+                                                                .addComponent(jLabel_enemy_level)))
+                                                .addGap(18, 18, 18)
+                                                .addComponent(jLabel_enemy_name, javax.swing.GroupLayout.PREFERRED_SIZE, 29, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                .addGroup(jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
+                                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                                .addGap(16, 16, 16)
+                                                                .addComponent(jLabel_self_level)
+                                                                .addGap(7, 7, 7)
+                                                                .addComponent(jLabel_self_image, javax.swing.GroupLayout.PREFERRED_SIZE, 227, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                                        .addGroup(jPanel_fightLayout.createSequentialGroup()
+                                                                .addGap(33, 33, 33)
+                                                                .addComponent(jLabel_what_happened, javax.swing.GroupLayout.PREFERRED_SIZE, 35, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                                                .addComponent(jLabel_turn, javax.swing.GroupLayout.PREFERRED_SIZE, 43, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                                .addGap(18, 18, 18)
+                                                                .addComponent(jLabel29, javax.swing.GroupLayout.PREFERRED_SIZE, 38, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                                                .addGap(18, 18, 18)
+                                                .addComponent(jLabel_Kitana, javax.swing.GroupLayout.PREFERRED_SIZE, 24, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 25, Short.MAX_VALUE)
+                                                .addGroup(jPanel_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                                                        .addComponent(jButton_items, javax.swing.GroupLayout.DEFAULT_SIZE, 44, Short.MAX_VALUE)
+                                                        .addComponent(jButton_attack, javax.swing.GroupLayout.DEFAULT_SIZE, 44, Short.MAX_VALUE)
+                                                        .addComponent(jButton_defend, javax.swing.GroupLayout.DEFAULT_SIZE, 44, Short.MAX_VALUE))
+                                                .addGap(14, 14, 14))))
+        );
+
+        javax.swing.GroupLayout jFrame_fightLayout = new javax.swing.GroupLayout(jFrame_fight.getContentPane());
+        jFrame_fight.getContentPane().setLayout(jFrame_fightLayout);
+        jFrame_fightLayout.setHorizontalGroup(
+                jFrame_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jFrame_fightLayout.createSequentialGroup()
+                                .addComponent(jPanel_fight, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                .addContainerGap())
+        );
+        jFrame_fightLayout.setVerticalGroup(
+                jFrame_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addComponent(jPanel_fight, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+
+        jPanel_end_of_fight.setBackground(new java.awt.Color(179, 226, 217));
+        jLabel_end_of_fight.setBackground(new java.awt.Color(204, 204, 204));
+        jLabel_end_of_fight.setFont(new java.awt.Font("Comic Sans MS", 1, 42)); // NOI18N
+        jLabel_end_of_fight.setForeground(new java.awt.Color(204, 0, 0));
+        jLabel_end_of_fight.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        jLabel_end_of_fight.setText("jLabel18");
+        jLabel_end_of_fight.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
+
+        jButton_continue.setBackground(new java.awt.Color(114, 218, 142));
+        jButton_continue.setFont(new java.awt.Font("Comic Sans MS", 0, 24)); // NOI18N
+        jButton_continue.setText("Дальше");
+        jButton_continue.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton_continueActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout jPanel_end_of_fightLayout = new javax.swing.GroupLayout(jPanel_end_of_fight);
+        jPanel_end_of_fight.setLayout(jPanel_end_of_fightLayout);
+        jPanel_end_of_fightLayout.setHorizontalGroup(
+                jPanel_end_of_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jPanel_end_of_fightLayout.createSequentialGroup()
+                                .addGroup(jPanel_end_of_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                        .addGroup(jPanel_end_of_fightLayout.createSequentialGroup()
+                                                .addGap(208, 208, 208)
+                                                .addComponent(jButton_continue, javax.swing.GroupLayout.PREFERRED_SIZE, 241, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                        .addGroup(jPanel_end_of_fightLayout.createSequentialGroup()
+                                                .addGap(104, 104, 104)
+                                                .addComponent(jLabel_end_of_fight, javax.swing.GroupLayout.PREFERRED_SIZE, 467, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                                .addContainerGap(106, Short.MAX_VALUE))
+        );
+        jPanel_end_of_fightLayout.setVerticalGroup(
+                jPanel_end_of_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jPanel_end_of_fightLayout.createSequentialGroup()
+                                .addGap(89, 89, 89)
+                                .addComponent(jLabel_end_of_fight, javax.swing.GroupLayout.PREFERRED_SIZE, 143, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 108, Short.MAX_VALUE)
+                                .addComponent(jButton_continue, javax.swing.GroupLayout.PREFERRED_SIZE, 139, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(72, 72, 72))
+        );
+        Dimension dim = Toolkit.getDefaultToolkit().getScreenSize();
+        jDialog_end_of_fight.setLocation(dim.width/2-jDialog_end_of_fight.getSize().width/2, dim.height/2-jDialog_end_of_fight.getSize().height/2);
+        javax.swing.GroupLayout jDialog_end_of_fightLayout = new javax.swing.GroupLayout(jDialog_end_of_fight.getContentPane());
+        jDialog_end_of_fight.getContentPane().setLayout(jDialog_end_of_fightLayout);
+        jDialog_end_of_fightLayout.setHorizontalGroup(
+                jDialog_end_of_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jDialog_end_of_fightLayout.createSequentialGroup()
+                                .addContainerGap()
+                                .addComponent(jPanel_end_of_fight, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                .addContainerGap())
+        );
+        jDialog_end_of_fightLayout.setVerticalGroup(
+                jDialog_end_of_fightLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jDialog_end_of_fightLayout.createSequentialGroup()
+                                .addContainerGap()
+                                .addComponent(jPanel_end_of_fight, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                .addContainerGap())
+        );
+
+        jPanel_win_in_top.setBackground(new java.awt.Color(255, 204, 255));
+
+        jLabel_win_message_in_top.setFont(new java.awt.Font("Comic Sans MS", 3, 24)); // NOI18N
+        jLabel_win_message_in_top.setForeground(new java.awt.Color(255, 0, 0));
+        jLabel_win_message_in_top.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        jLabel_win_message_in_top.setText("Победа на вашей стороне");
+
+        jTextField_win_in_top.setFont(new java.awt.Font("Comic Sans MS", 1, 18)); // NOI18N
+        jTextField_win_in_top.setHorizontalAlignment(javax.swing.JTextField.CENTER);
+        jTextField_win_in_top.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+            }
+        });
+
+        jLabel_in_top_message1.setFont(new java.awt.Font("Comic Sans MS", 0, 14)); // NOI18N
+        jLabel_in_top_message1.setForeground(new java.awt.Color(102, 102, 102));
+        jLabel_in_top_message1.setText("Введите имя своего персонажа для добавления");
+
+        jLabel_in_top_message2.setFont(new java.awt.Font("Comic Sans MS", 0, 14)); // NOI18N
+        jLabel_in_top_message2.setForeground(new java.awt.Color(102, 102, 102));
+        jLabel_in_top_message2.setText("результата в таблицу рекордов");
+
+        jButton_finish_game_in_top.setBackground(new java.awt.Color(153, 153, 255));
+        jButton_finish_game_in_top.setFont(new java.awt.Font("Comic Sans MS", 1, 14)); // NOI18N
+        jButton_finish_game_in_top.setForeground(new java.awt.Color(51, 51, 51));
+        jButton_finish_game_in_top.setText("Закончить игру");
+        jButton_finish_game_in_top.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton_finish_game_in_topActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout jPanel_win_in_topLayout = new javax.swing.GroupLayout(jPanel_win_in_top);
+        jPanel_win_in_top.setLayout(jPanel_win_in_topLayout);
+        jPanel_win_in_topLayout.setHorizontalGroup(
+                jPanel_win_in_topLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel_win_in_topLayout.createSequentialGroup()
+                                .addGap(0, 0, Short.MAX_VALUE)
+                                .addComponent(jButton_finish_game_in_top)
+                                .addGap(14, 14, 14))
+                        .addGroup(jPanel_win_in_topLayout.createSequentialGroup()
+                                .addGroup(jPanel_win_in_topLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                        .addGroup(jPanel_win_in_topLayout.createSequentialGroup()
+                                                .addGap(132, 132, 132)
+                                                .addComponent(jLabel_in_top_message2))
+                                        .addGroup(jPanel_win_in_topLayout.createSequentialGroup()
+                                                .addGap(60, 60, 60)
+                                                .addComponent(jLabel_win_message_in_top, javax.swing.GroupLayout.PREFERRED_SIZE, 363, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                        .addGroup(jPanel_win_in_topLayout.createSequentialGroup()
+                                                .addGap(77, 77, 77)
+                                                .addGroup(jPanel_win_in_topLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                                        .addComponent(jTextField_win_in_top, javax.swing.GroupLayout.PREFERRED_SIZE, 312, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                        .addComponent(jLabel_in_top_message1))))
+                                .addContainerGap(62, Short.MAX_VALUE))
+        );
+        jPanel_win_in_topLayout.setVerticalGroup(
+                jPanel_win_in_topLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jPanel_win_in_topLayout.createSequentialGroup()
+                                .addGap(49, 49, 49)
+                                .addComponent(jLabel_win_message_in_top, javax.swing.GroupLayout.PREFERRED_SIZE, 81, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                                .addComponent(jLabel_in_top_message1, javax.swing.GroupLayout.PREFERRED_SIZE, 38, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                .addComponent(jLabel_in_top_message2)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 33, Short.MAX_VALUE)
+                                .addComponent(jTextField_win_in_top, javax.swing.GroupLayout.PREFERRED_SIZE, 56, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                .addComponent(jButton_finish_game_in_top, javax.swing.GroupLayout.PREFERRED_SIZE, 50, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addContainerGap())
+        );
+
+        javax.swing.GroupLayout jDialog_win_in_topLayout = new javax.swing.GroupLayout(jDialog_win_in_top.getContentPane());
+        jDialog_win_in_top.getContentPane().setLayout(jDialog_win_in_topLayout);
+        jDialog_win_in_topLayout.setHorizontalGroup(
+                jDialog_win_in_topLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jDialog_win_in_topLayout.createSequentialGroup()
+                                .addContainerGap()
+                                .addComponent(jPanel_win_in_top, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                .addContainerGap())
+        );
+        jDialog_win_in_topLayout.setVerticalGroup(
+                jDialog_win_in_topLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jDialog_win_in_topLayout.createSequentialGroup()
+                                .addContainerGap()
+                                .addComponent(jPanel_win_in_top, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                .addContainerGap())
+        );
+
+        jPanel_records_table.setBackground(new java.awt.Color(204, 204, 255));
+
+        jLabel_records_table.setFont(new java.awt.Font("Comic Sans MS", 1, 24)); // NOI18N
+        jLabel_records_table.setForeground(new java.awt.Color(51, 51, 51));
+        jLabel_records_table.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        jLabel_records_table.setText("Таблица рекордов");
+
+        jTable_records.setModel(new javax.swing.table.DefaultTableModel(
+                new Object [][] {
+                        {null, null},
+                        {null, null},
+                        {null, null},
+                        {null, null},
+                        {null, null},
+                        {null, null},
+                        {null, null},
+                        {null, null},
+                        {null, null},
+                        {null, null}
+                },
+                new String [] {
+                        "Имя", "Количество баллов"
+                }
+        ) {
+            Class[] types = new Class [] {
+                    java.lang.String.class, java.lang.Integer.class
+            };
+
+            public Class getColumnClass(int columnIndex) {
+                return types [columnIndex];
+            }
+        });
+        jScrollPane_records_table.setViewportView(jTable_records);
+
+        jButton_close_records_table.setBackground(new java.awt.Color(255, 255, 153));
+        jButton_close_records_table.setFont(new java.awt.Font("Comic Sans MS", 0, 18)); // NOI18N
+        jButton_close_records_table.setForeground(new java.awt.Color(51, 51, 51));
+        jButton_close_records_table.setText("Закрыть");
+        jButton_close_records_table.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton_close_records_tableActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout jPanel_records_tableLayout = new javax.swing.GroupLayout(jPanel_records_table);
+        jPanel_records_table.setLayout(jPanel_records_tableLayout);
+        jPanel_records_tableLayout.setHorizontalGroup(
+                jPanel_records_tableLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jPanel_records_tableLayout.createSequentialGroup()
+                                .addGroup(jPanel_records_tableLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                                        .addComponent(jButton_close_records_table, javax.swing.GroupLayout.PREFERRED_SIZE, 143, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                        .addGroup(jPanel_records_tableLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                                .addGroup(jPanel_records_tableLayout.createSequentialGroup()
+                                                        .addGap(160, 160, 160)
+                                                        .addComponent(jLabel_records_table, javax.swing.GroupLayout.PREFERRED_SIZE, 241, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                                .addGroup(jPanel_records_tableLayout.createSequentialGroup()
+                                                        .addGap(18, 18, 18)
+                                                        .addComponent(jScrollPane_records_table, javax.swing.GroupLayout.PREFERRED_SIZE, 527, javax.swing.GroupLayout.PREFERRED_SIZE))))
+                                .addContainerGap(17, Short.MAX_VALUE))
+        );
+        jPanel_records_tableLayout.setVerticalGroup(
+                jPanel_records_tableLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jPanel_records_tableLayout.createSequentialGroup()
+                                .addGap(15, 15, 15)
+                                .addComponent(jLabel_records_table, javax.swing.GroupLayout.PREFERRED_SIZE, 43, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                .addComponent(jScrollPane_records_table, javax.swing.GroupLayout.PREFERRED_SIZE, 262, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                                .addComponent(jButton_close_records_table, javax.swing.GroupLayout.DEFAULT_SIZE, 38, Short.MAX_VALUE)
+                                .addContainerGap())
+        );
+
+        javax.swing.GroupLayout jDialog_records_tableLayout = new javax.swing.GroupLayout(jDialog_records_table.getContentPane());
+        jDialog_records_table.getContentPane().setLayout(jDialog_records_tableLayout);
+        jDialog_records_tableLayout.setHorizontalGroup(
+                jDialog_records_tableLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addComponent(jPanel_records_table, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+        jDialog_records_tableLayout.setVerticalGroup(
+                jDialog_records_tableLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addComponent(jPanel_records_table, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+
+        jPanel_win_outside_top.setBackground(new java.awt.Color(255, 204, 255));
+
+        jLabel_win_message_outside_top.setFont(new java.awt.Font("Comic Sans MS", 3, 24)); // NOI18N
+        jLabel_win_message_outside_top.setForeground(new java.awt.Color(255, 0, 0));
+        jLabel_win_message_outside_top.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        jLabel_win_message_outside_top.setText("Победа на вашей стороне");
+
+        jLabel_outside_top_message.setFont(new java.awt.Font("Comic Sans MS", 0, 14)); // NOI18N
+        jLabel_outside_top_message.setForeground(new java.awt.Color(102, 102, 102));
+        jLabel_outside_top_message.setText("К сожалению, Ваш результат не попал в топ 10");
+
+        jButton_finish_game_outside_top.setBackground(new java.awt.Color(153, 153, 255));
+        jButton_finish_game_outside_top.setFont(new java.awt.Font("Comic Sans MS", 1, 14)); // NOI18N
+        jButton_finish_game_outside_top.setForeground(new java.awt.Color(51, 51, 51));
+        jButton_finish_game_outside_top.setText("Закончить игру");
+        jButton_finish_game_outside_top.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton_finish_game_outside_topActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout jPanel_win_outside_topLayout = new javax.swing.GroupLayout(jPanel_win_outside_top);
+        jPanel_win_outside_top.setLayout(jPanel_win_outside_topLayout);
+        jPanel_win_outside_topLayout.setHorizontalGroup(
+                jPanel_win_outside_topLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jPanel_win_outside_topLayout.createSequentialGroup()
+                                .addGroup(jPanel_win_outside_topLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                        .addGroup(jPanel_win_outside_topLayout.createSequentialGroup()
+                                                .addGap(48, 48, 48)
+                                                .addComponent(jLabel_win_message_outside_top, javax.swing.GroupLayout.PREFERRED_SIZE, 354, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                        .addGroup(jPanel_win_outside_topLayout.createSequentialGroup()
+                                                .addGap(58, 58, 58)
+                                                .addComponent(jLabel_outside_top_message, javax.swing.GroupLayout.PREFERRED_SIZE, 333, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                        .addGroup(jPanel_win_outside_topLayout.createSequentialGroup()
+                                                .addGap(153, 153, 153)
+                                                .addComponent(jButton_finish_game_outside_top, javax.swing.GroupLayout.PREFERRED_SIZE, 158, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                                .addContainerGap(68, Short.MAX_VALUE))
+        );
+        jPanel_win_outside_topLayout.setVerticalGroup(
+                jPanel_win_outside_topLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jPanel_win_outside_topLayout.createSequentialGroup()
+                                .addGap(60, 60, 60)
+                                .addComponent(jLabel_win_message_outside_top, javax.swing.GroupLayout.PREFERRED_SIZE, 73, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(41, 41, 41)
+                                .addComponent(jLabel_outside_top_message, javax.swing.GroupLayout.PREFERRED_SIZE, 29, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(56, 56, 56)
+                                .addComponent(jButton_finish_game_outside_top, javax.swing.GroupLayout.PREFERRED_SIZE, 51, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addContainerGap(56, Short.MAX_VALUE))
+        );
+
+        javax.swing.GroupLayout jDialog_win_outside_topLayout = new javax.swing.GroupLayout(jDialog_win_outside_top.getContentPane());
+        jDialog_win_outside_top.getContentPane().setLayout(jDialog_win_outside_topLayout);
+        jDialog_win_outside_topLayout.setHorizontalGroup(
+                jDialog_win_outside_topLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addComponent(jPanel_win_outside_top, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+        jDialog_win_outside_topLayout.setVerticalGroup(
+                jDialog_win_outside_topLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addComponent(jPanel_win_outside_top, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+
+        jPanel_bag_with_items.setBackground(new java.awt.Color(190, 182, 135));
+
+        jLabel_bag_of_items.setFont(new java.awt.Font("Comic Sans MS", 0, 16)); // NOI18N
+        jLabel_bag_of_items.setForeground(new java.awt.Color(51, 51, 51));
+        jLabel_bag_of_items.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        jLabel_bag_of_items.setText("Мешок предметов");
+
+        jRadioButton_small_potion.setBackground(new java.awt.Color(190, 182, 135));
+        jRadioButton_small_potion.setFont(new java.awt.Font("Comic Sans MS", 0, 12)); // NOI18N
+        jRadioButton_small_potion.setText("Малое зелье лечения, 0 шт");
+        jRadioButton_small_potion.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {}
+        });
+
+        jRadioButton_big_potion.setBackground(new java.awt.Color(190, 182, 135));
+        jRadioButton_big_potion.setFont(new java.awt.Font("Comic Sans MS", 0, 12)); // NOI18N
+        jRadioButton_big_potion.setText("Большое зелье лечения, 0 шт");
+        jRadioButton_big_potion.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {}
+        });
+
+        jRadioButton_resurrection_cross.setBackground(new java.awt.Color(190, 182, 135));
+        jRadioButton_resurrection_cross.setFont(new java.awt.Font("Comic Sans MS", 0, 12)); // NOI18N
+        jRadioButton_resurrection_cross.setText("Крест возрождения, 0 шт");
+
+        jButton_use_item.setBackground(new java.awt.Color(239, 237, 14));
+        jButton_use_item.setFont(new java.awt.Font("Comic Sans MS", 1, 14)); // NOI18N
+        jButton_use_item.setText("Использовать");
+        jButton_use_item.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton_use_itemActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout jPanel_bag_with_itemsLayout = new javax.swing.GroupLayout(jPanel_bag_with_items);
+        jPanel_bag_with_items.setLayout(jPanel_bag_with_itemsLayout);
+        jPanel_bag_with_itemsLayout.setHorizontalGroup(
+                jPanel_bag_with_itemsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jPanel_bag_with_itemsLayout.createSequentialGroup()
+                                .addGroup(jPanel_bag_with_itemsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                        .addGroup(jPanel_bag_with_itemsLayout.createSequentialGroup()
+                                                .addGap(120, 120, 120)
+                                                .addGroup(jPanel_bag_with_itemsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                                                        .addComponent(jRadioButton_big_potion)
+                                                        .addComponent(jRadioButton_resurrection_cross, javax.swing.GroupLayout.Alignment.LEADING)
+                                                        .addComponent(jRadioButton_small_potion, javax.swing.GroupLayout.Alignment.LEADING)
+                                                        .addComponent(jLabel_bag_of_items, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.PREFERRED_SIZE, 168, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                                        .addGroup(jPanel_bag_with_itemsLayout.createSequentialGroup()
+                                                .addGap(139, 139, 139)
+                                                .addComponent(jButton_use_item)))
+                                .addContainerGap(105, Short.MAX_VALUE))
+        );
+        jPanel_bag_with_itemsLayout.setVerticalGroup(
+                jPanel_bag_with_itemsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jPanel_bag_with_itemsLayout.createSequentialGroup()
+                                .addGap(30, 30, 30)
+                                .addComponent(jLabel_bag_of_items, javax.swing.GroupLayout.PREFERRED_SIZE, 47, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(40, 40, 40)
+                                .addComponent(jRadioButton_small_potion)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                .addComponent(jRadioButton_big_potion)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                .addComponent(jRadioButton_resurrection_cross)
+                                .addGap(50, 50, 50)
+                                .addComponent(jButton_use_item, javax.swing.GroupLayout.PREFERRED_SIZE, 38, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addContainerGap(54, Short.MAX_VALUE))
+        );
+
+        javax.swing.GroupLayout jDialog_bag_with_itemsLayout = new javax.swing.GroupLayout(jDialog_bag_with_items.getContentPane());
+        jDialog_bag_with_items.getContentPane().setLayout(jDialog_bag_with_itemsLayout);
+        jDialog_bag_with_itemsLayout.setHorizontalGroup(
+                jDialog_bag_with_itemsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addComponent(jPanel_bag_with_items, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+        jDialog_bag_with_itemsLayout.setVerticalGroup(
+                jDialog_bag_with_itemsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addComponent(jPanel_bag_with_items, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+
+        jPanel_unavailable_item.setBackground(new java.awt.Color(243, 120, 120));
+
+        jLabel_unavailable_item1.setFont(new java.awt.Font("Comic Sans MS", 1, 18)); // NOI18N
+        jLabel_unavailable_item1.setForeground(new java.awt.Color(51, 51, 51));
+        jLabel_unavailable_item1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        jLabel_unavailable_item1.setText("Вы не можете использовать ");
+
+        jLabel_unavailable_item2.setFont(new java.awt.Font("Comic Sans MS", 1, 18)); // NOI18N
+        jLabel_unavailable_item2.setForeground(new java.awt.Color(51, 51, 51));
+        jLabel_unavailable_item2.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        jLabel_unavailable_item2.setText("этот предмет");
+
+        jButton_unavailable_item.setBackground(new java.awt.Color(204, 204, 204));
+        jButton_unavailable_item.setFont(new java.awt.Font("Comic Sans MS", 0, 18)); // NOI18N
+        jButton_unavailable_item.setForeground(new java.awt.Color(51, 51, 51));
+        jButton_unavailable_item.setText("ОК");
+        jButton_unavailable_item.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton_unavailable_itemActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout jPanel_unavailable_itemLayout = new javax.swing.GroupLayout(jPanel_unavailable_item);
+        jPanel_unavailable_item.setLayout(jPanel_unavailable_itemLayout);
+        jPanel_unavailable_itemLayout.setHorizontalGroup(
+                jPanel_unavailable_itemLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jPanel_unavailable_itemLayout.createSequentialGroup()
+                                .addContainerGap(82, Short.MAX_VALUE)
+                                .addGroup(jPanel_unavailable_itemLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                        .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel_unavailable_itemLayout.createSequentialGroup()
+                                                .addComponent(jLabel_unavailable_item1)
+                                                .addGap(58, 58, 58))
+                                        .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel_unavailable_itemLayout.createSequentialGroup()
+                                                .addComponent(jLabel_unavailable_item2, javax.swing.GroupLayout.PREFERRED_SIZE, 126, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                .addGap(135, 135, 135))))
+                        .addGroup(jPanel_unavailable_itemLayout.createSequentialGroup()
+                                .addGap(150, 150, 150)
+                                .addComponent(jButton_unavailable_item, javax.swing.GroupLayout.PREFERRED_SIZE, 101, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(0, 0, Short.MAX_VALUE))
+        );
+        jPanel_unavailable_itemLayout.setVerticalGroup(
+                jPanel_unavailable_itemLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jPanel_unavailable_itemLayout.createSequentialGroup()
+                                .addGap(59, 59, 59)
+                                .addComponent(jLabel_unavailable_item1, javax.swing.GroupLayout.PREFERRED_SIZE, 58, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                .addComponent(jLabel_unavailable_item2)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 62, Short.MAX_VALUE)
+                                .addComponent(jButton_unavailable_item, javax.swing.GroupLayout.PREFERRED_SIZE, 37, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(52, 52, 52))
+        );
+
+        javax.swing.GroupLayout jDialog_unavailable_itemLayout = new javax.swing.GroupLayout(jDialog_unavailable_item.getContentPane());
+        jDialog_unavailable_item.getContentPane().setLayout(jDialog_unavailable_itemLayout);
+        jDialog_unavailable_itemLayout.setHorizontalGroup(
+                jDialog_unavailable_itemLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addComponent(jPanel_unavailable_item, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+        jDialog_unavailable_itemLayout.setVerticalGroup(
+                jDialog_unavailable_itemLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addComponent(jPanel_unavailable_item, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+
+        jLabel_locations.setFont(new java.awt.Font("Segoe UI", 0, 14)); // NOI18N
+        jLabel_locations.setText("Введите количество локаций");
+
+        jButton_locations_start.setText("Старт");
+        jButton_locations_start.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton_locations_startActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout jPanel_locationsLayout = new javax.swing.GroupLayout(jPanel_locations);
+        jPanel_locations.setLayout(jPanel_locationsLayout);
+
+        jPanel_locationsLayout.setHorizontalGroup(
+                jPanel_locationsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel_locationsLayout.createSequentialGroup()
+                                .addContainerGap(110, Short.MAX_VALUE)
+                                .addComponent(jLabel_locations)
+                                .addGap(106, 106, 106))
+                        .addGroup(jPanel_locationsLayout.createSequentialGroup()
+                                .addGap(202, 202, 202)
+                                .addGroup(jPanel_locationsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                        .addComponent(jButton_locations_start)
+                                        .addComponent(jTextField_locations, javax.swing.GroupLayout.PREFERRED_SIZE, 71, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+        );
+        jPanel_locationsLayout.setVerticalGroup(
+                jPanel_locationsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jPanel_locationsLayout.createSequentialGroup()
+                                .addGap(22, 22, 22)
+                                .addComponent(jLabel_locations, javax.swing.GroupLayout.PREFERRED_SIZE, 84, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                .addComponent(jTextField_locations, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(28, 28, 28)
+                                .addComponent(jButton_locations_start)
+                                .addContainerGap(101, Short.MAX_VALUE))
+        );
+
+        javax.swing.GroupLayout jDialog_locationsLayout = new javax.swing.GroupLayout(jDialog_locations.getContentPane());
+        jDialog_locations.getContentPane().setLayout(jDialog_locationsLayout);
+        jDialog_locationsLayout.setHorizontalGroup(
+                jDialog_locationsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGap(0, 481, Short.MAX_VALUE)
+                        .addGroup(jDialog_locationsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                .addGroup(jDialog_locationsLayout.createSequentialGroup()
+                                        .addGap(0, 0, Short.MAX_VALUE)
+                                        .addComponent(jPanel_locations, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                        .addGap(0, 0, Short.MAX_VALUE)))
+        );
+        jDialog_locationsLayout.setVerticalGroup(
+                jDialog_locationsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGap(0, 300, Short.MAX_VALUE)
+                        .addGroup(jDialog_locationsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                .addGroup(jDialog_locationsLayout.createSequentialGroup()
+                                        .addGap(0, 0, Short.MAX_VALUE)
+                                        .addComponent(jPanel_locations, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                        .addGap(0, 0, Short.MAX_VALUE)))
+        );
+
+        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
+
+        jPanel_start.setBackground(new java.awt.Color(255, 255, 255));
+
+        jLabel_mortal_kombat.setFont(new java.awt.Font("Comic Sans MS", 2, 24)); // NOI18N
+        jLabel_mortal_kombat.setForeground(new java.awt.Color(204, 0, 30));
+        jLabel_mortal_kombat.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        jLabel_mortal_kombat.setText("Mortal Kombat");
+
+        jButton_start_game.setFont(new java.awt.Font("Comic Sans MS", 0, 12)); // NOI18N
+        jButton_start_game.setText("Начать новую игру");
+        jButton_start_game.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton_start_gameActionPerformed(evt);
+            }
+        });
+
+        jButton_watch_results.setFont(new java.awt.Font("Comic Sans MS", 0, 12)); // NOI18N
+        jButton_watch_results.setText("Посмотреть таблицу \nрезультатов");
+        jButton_watch_results.setHorizontalAlignment(javax.swing.SwingConstants.TRAILING);
+        jButton_watch_results.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton_watch_resultsActionPerformed(evt);
+            }
+        });
+
+        URL url = Objects.requireNonNull(getClass().getClassLoader().getResource("img/mk.png"));
+        try {
+            BufferedImage originalImage = ImageIO.read(url);
+            // Изменение размера изображения до 300x300 пикселей
+            BufferedImage resizedImage = resizeImage(originalImage, 300, 300);
+
+            // Установка измененного изображения в JLabel
+            jLabel3 = new JLabel(new ImageIcon(resizedImage));
+        } catch (Exception e) {
+            System.out.println("img mk is not found");
+        }
+
+
+        javax.swing.GroupLayout jPanel_startLayout = new javax.swing.GroupLayout(jPanel_start);
+        jPanel_start.setLayout(jPanel_startLayout);
+        jPanel_startLayout.setHorizontalGroup(
+                jPanel_startLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jPanel_startLayout.createSequentialGroup()
+                                .addGap(22, 22, 22)
+                                .addComponent(jButton_watch_results)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 43, Short.MAX_VALUE)
+                                .addComponent(jButton_start_game, javax.swing.GroupLayout.PREFERRED_SIZE, 214, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(21, 21, 21))
+                        .addGroup(jPanel_startLayout.createSequentialGroup()
+                                .addGroup(jPanel_startLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                        .addGroup(jPanel_startLayout.createSequentialGroup()
+                                                .addGap(159, 159, 159)
+                                                .addComponent(jLabel_mortal_kombat, javax.swing.GroupLayout.PREFERRED_SIZE, 174, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                        .addGroup(jPanel_startLayout.createSequentialGroup()
+                                                .addGap(143, 143, 143)
+                                                .addComponent(jLabel3)))
+                                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+        );
+        jPanel_startLayout.setVerticalGroup(
+                jPanel_startLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(jPanel_startLayout.createSequentialGroup()
+                                .addGap(54, 54, 54)
+                                .addComponent(jLabel_mortal_kombat, javax.swing.GroupLayout.PREFERRED_SIZE, 44, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                                .addComponent(jLabel3)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 40, Short.MAX_VALUE)
+                                .addGroup(jPanel_startLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                                        .addComponent(jButton_watch_results, javax.swing.GroupLayout.PREFERRED_SIZE, 70, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                        .addComponent(jButton_start_game, javax.swing.GroupLayout.PREFERRED_SIZE, 70, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                .addGap(30, 30, 30))
+        );
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
+        getContentPane().setLayout(layout);
+        layout.setHorizontalGroup(
+                layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(layout.createSequentialGroup()
+                                .addContainerGap()
+                                .addComponent(jPanel_start, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                .addContainerGap())
+        );
+        layout.setVerticalGroup(
+                layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addGroup(layout.createSequentialGroup()
+                                .addContainerGap()
+                                .addComponent(jPanel_start, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                .addContainerGap())
+        );
+
+        pack();
+    }// </editor-fold>//GEN-END:initComponents
+    /**
+     * Кнопка начала игры
+     */
+    private void jButton_start_gameActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton_start_gameActionPerformed
+        jDialog_locations.setVisible(true);
+        jDialog_locations.setBounds(200, 100, 500, 300);
+        Dimension dim = Toolkit.getDefaultToolkit().getScreenSize();
+        jDialog_locations.setLocation(dim.width/2-jDialog_locations.getSize().width/2, dim.height/2-jDialog_locations.getSize().height/2);
+    }//GEN-LAST:event_jButton_start_gameActionPerformed
+    /**
+     * Кнопка атаки
+     */
+    private void jButton_attackActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton_attackActionPerformed
+        oldGame.fight.hit(human, enemy, 1, jLabel_enemy_health, jLabel_self_health, jDialog_end_of_fight, jLabel_end_of_fight, oldGame.action,
+                jProgressBar_self, jProgressBar_enemy, jDialog_win_in_top, jDialog_win_outside_top, jFrame_fight,
+                oldGame.getResults(), jLabel_win_message_in_top, jLabel_win_message_outside_top, jLabel_turn, jLabel29,
+                jLabel_what_happened, items, jRadioButton_resurrection_cross,  jLabel_location, jPanel_fight);
+    }//GEN-LAST:event_jButton_attackActionPerformed
+    /**
+     * Кнопка защиты
+     */
+    private void jButton_defendActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton4_defendActionPerformed
+        oldGame.fight.hit(human, enemy, 0, jLabel_enemy_health, jLabel_self_health, jDialog_end_of_fight, jLabel_end_of_fight, oldGame.action,
+                jProgressBar_self, jProgressBar_enemy, jDialog_win_in_top, jDialog_win_outside_top, jFrame_fight,
+                oldGame.getResults(), jLabel_win_message_in_top, jLabel_win_message_outside_top, jLabel_turn, jLabel29,
+                jLabel_what_happened, items, jRadioButton_resurrection_cross,  jLabel_location, jPanel_fight);
+    }//GEN-LAST:event_jButton4_defendActionPerformed
+    /**
+     * Кнопка продолжения игры после завершения раунда
+     */
+    private void jButton_continueActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton_continueActionPerformed
+        enemy = oldGame.fight.newRound(human, jLabel_enemy_image, jProgressBar_self, jProgressBar_enemy,
+                jLabel_enemy_name, jLabel_enemy_damage_value, jLabel_enemy_health, oldGame.action);
+        oldGame.change.newRoundTexts(human, enemy, jProgressBar_self, jProgressBar_enemy, jLabel_points_value, jLabel_experience_value,
+                jLabel_self_level, jLabel_enemy_level, jLabel_self_health, jLabel_enemy_health, jLabel_self_damage_value,
+                jLabel_turn, jLabel_what_happened, oldGame.fight.i, items, jRadioButton_small_potion, jRadioButton_big_potion, jRadioButton_resurrection_cross);
+        jDialog_end_of_fight.dispose();
+    }//GEN-LAST:event_jButton_continueActionPerformed
+    /**
+     * Кнопка завершения игры с результатом в топ-10
+     */
+    private void jButton_finish_game_in_topActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton_finish_game_in_topActionPerformed
+        try {
+            oldGame.endGameTop(human, jTextField_win_in_top, jTable_records);
+        } catch (IOException ex) {
+            Logger.getLogger(JFrames.class.getName()).log(Level.SEVERE, null, ex);
+        }
+        jDialog_win_in_top.dispose();
+        jTextField_win_in_top.setText("");
+    }//GEN-LAST:event_jButton_finish_game_in_topActionPerformed
+    /**
+     * Кнопка закрытия таблицы рейтинга
+     */
+    private void jButton_close_records_tableActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton_close_records_tableActionPerformed
+        jDialog_records_table.dispose();
+    }//GEN-LAST:event_jButton_close_records_tableActionPerformed
+    /**
+     * Кнопка открытия таблицы с рейтингом
+     */
+    private void jButton_watch_resultsActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton_watch_resultsActionPerformed
+        jDialog_records_table.setVisible(true);
+        jDialog_records_table.setBounds(100, 100, 580, 450);
+    }//GEN-LAST:event_jButton_watch_resultsActionPerformed
+    /**
+     * Кнопка завершения игры при непопадании в топ-10
+     */
+    private void jButton_finish_game_outside_topActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton_finish_game_outside_topActionPerformed
+        jDialog_win_outside_top.dispose();
+    }//GEN-LAST:event_jButton_finish_game_outside_topActionPerformed
+    /**
+     * Кнопка использования предметов
+     */
+    private void jButton_use_itemActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton_use_itemActionPerformed
+        String nameButton = "";
+        if(jRadioButton_small_potion.isSelected()){
+            nameButton = "jRadioButton1";
+        }
+        if(jRadioButton_big_potion.isSelected()){
+            nameButton = "jRadioButton2";
+        }
+        if(jRadioButton_resurrection_cross.isSelected()){
+            nameButton = "jRadioButton3";
+        }
+        oldGame.action.useItem(human, items, nameButton, jDialog_unavailable_item, jDialog_bag_with_items);
+        oldGame.action.hp(human, jProgressBar_self);
+        jLabel_self_health.setText(human.getHealth() + "/" + human.getMaxHealth());
+        oldGame.change.BagText(items, jRadioButton_small_potion, jRadioButton_big_potion, jRadioButton_resurrection_cross);
+    }//GEN-LAST:event_jButton_use_itemActionPerformed
+    /**
+     * Кнопка получения списка доступных предметов
+     */
+    private void jButton_itemsActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton_itemsActionPerformed
+        jDialog_bag_with_items.setVisible(true);
+        jDialog_bag_with_items.setBounds(300, 200, 430, 350);
+    }//GEN-LAST:event_jButton_itemsActionPerformed
+    /**
+     * Кнопка закрытия окна с уведомлением о невозможности использования предмета
+     */
+    private void jButton_unavailable_itemActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton_unavailable_itemActionPerformed
+        jDialog_unavailable_item.dispose();
+    }//GEN-LAST:event_jButton_unavailable_itemActionPerformed
+    /**
+     * Кнопка начала игры после выбора количества локаций
+     */
+    private void jButton_locations_startActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton_locations_startActionPerformed
+        int n_locations;
+        try {
+            n_locations = Integer.parseInt(jTextField_locations.getText());
+            if (n_locations <= 0 || n_locations > 10) {
+                throw new Exception();
+            }
+        } catch (Exception e) {
+            JOptionPane.showMessageDialog(null, "Invalid number of locations! Try again", "Error", JOptionPane.ERROR_MESSAGE);
+            jTextField_locations.setText("");
+        }
+        n_locations = Integer.parseInt(jTextField_locations.getText());
+        jDialog_locations.dispose();
+        jFrame_fight.setVisible(rootPaneCheckingEnabled);
+        jFrame_fight.setSize(1000, 800);
+        ImageIcon icon1 = new ImageIcon(getClass().getClassLoader().getResource("img/gamaz.jpg"));
+        icon1 = new ImageIcon(icon1.getImage().getScaledInstance(
+                200, 290, java.awt.Image.SCALE_SMOOTH));
+        jLabel_self_image.setIcon(icon1);
+        jLabel_turn.setText("Your turn");
+        //game.setLocations_number(Integer.parseInt(jTextField_locations.getText()));
+        human = oldGame.newHuman(jProgressBar_self, n_locations);
+        enemy = oldGame.newEnemy(jLabel_enemy_image, jLabel_enemy_name, jLabel_enemy_damage_value, jLabel_enemy_health, jProgressBar_enemy);
+        oldGame.change.newRoundTexts(human, enemy, jProgressBar_self, jProgressBar_enemy,
+                jLabel_points_value, jLabel_experience_value, jLabel_self_level, jLabel_enemy_level, jLabel_self_health, jLabel_enemy_health, jLabel_self_damage_value,
+                jLabel_turn, jLabel_what_happened, oldGame.fight.i, items, jRadioButton_small_potion, jRadioButton_big_potion, jRadioButton_resurrection_cross);
+    }//GEN-LAST:event_jButton_locations_startActionPerformed
+
+    /**
+     * @param args the command line arguments
+     */
+    public static void main(String args[]) {
+        /* Set the Nimbus look and feel */
+        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
+        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
+         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html
+         */
+        try {
+            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
+                if ("Nimbus".equals(info.getName())) {
+                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
+                    break;
+                }
+            }
+        } catch (ClassNotFoundException ex) {
+            Logger.getLogger(JFrames.class.getName()).log(Level.SEVERE, null, ex);
+        } catch (InstantiationException ex) {
+            Logger.getLogger(JFrames.class.getName()).log(Level.SEVERE, null, ex);
+        } catch (IllegalAccessException ex) {
+            Logger.getLogger(JFrames.class.getName()).log(Level.SEVERE, null, ex);
+        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
+            Logger.getLogger(JFrames.class.getName()).log(Level.SEVERE, null, ex);
+        }
+        //</editor-fold>
+
+        /* Create and display the form */
+        java.awt.EventQueue.invokeLater(new Runnable() {
+            public void run() {
+                new JFrames().setVisible(true);
+            }
+        });
+    }
+
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.ButtonGroup buttonGroup_items;
+    private javax.swing.JButton jButton_attack;
+    private javax.swing.JButton jButton_close_records_table;
+    private javax.swing.JButton jButton_continue;
+    private javax.swing.JButton jButton_defend;
+    private javax.swing.JButton jButton_finish_game_in_top;
+    private javax.swing.JButton jButton_finish_game_outside_top;
+    private javax.swing.JButton jButton_items;
+    private javax.swing.JButton jButton_locations_start;
+    private javax.swing.JButton jButton_start_game;
+    private javax.swing.JButton jButton_unavailable_item;
+    private javax.swing.JButton jButton_use_item;
+    private javax.swing.JButton jButton_watch_results;
+    private javax.swing.JDialog jDialog_bag_with_items;
+    private javax.swing.JDialog jDialog_end_of_fight;
+    private javax.swing.JDialog jDialog_locations;
+    private javax.swing.JDialog jDialog_records_table;
+    private javax.swing.JDialog jDialog_unavailable_item;
+    private javax.swing.JDialog jDialog_win_in_top;
+    private javax.swing.JDialog jDialog_win_outside_top;
+    private javax.swing.JFrame jFrame_fight;
+    private javax.swing.JLabel jLabel29;
+    private javax.swing.JLabel jLabel3;
+    private javax.swing.JLabel jLabel_Kitana;
+    private javax.swing.JLabel jLabel_bag_of_items;
+    private javax.swing.JLabel jLabel_end_of_fight;
+    private javax.swing.JLabel jLabel_enemy_damage;
+    private javax.swing.JLabel jLabel_enemy_damage_value;
+    private javax.swing.JLabel jLabel_enemy_health;
+    private javax.swing.JLabel jLabel_enemy_image;
+    private javax.swing.JLabel jLabel_enemy_level;
+    private javax.swing.JLabel jLabel_enemy_name;
+    private javax.swing.JLabel jLabel_experience;
+    private javax.swing.JLabel jLabel_experience_value;
+    private javax.swing.JLabel jLabel_fight;
+    private javax.swing.JLabel jLabel_in_top_message1;
+    private javax.swing.JLabel jLabel_in_top_message2;
+    private javax.swing.JLabel jLabel_location;
+    private javax.swing.JLabel jLabel_locations;
+    private javax.swing.JLabel jLabel_mortal_kombat;
+    private javax.swing.JLabel jLabel_outside_top_message;
+    private javax.swing.JLabel jLabel_points;
+    private javax.swing.JLabel jLabel_points_value;
+    private javax.swing.JLabel jLabel_records_table;
+    private javax.swing.JLabel jLabel_self_damage;
+    private javax.swing.JLabel jLabel_self_damage_value;
+    private javax.swing.JLabel jLabel_self_health;
+    private javax.swing.JLabel jLabel_self_image;
+    private javax.swing.JLabel jLabel_self_level;
+    private javax.swing.JLabel jLabel_turn;
+    private javax.swing.JLabel jLabel_unavailable_item1;
+    private javax.swing.JLabel jLabel_unavailable_item2;
+    private javax.swing.JLabel jLabel_what_happened;
+    private javax.swing.JLabel jLabel_win_message_in_top;
+    private javax.swing.JLabel jLabel_win_message_outside_top;
+    private javax.swing.JPanel jPanel_bag_with_items;
+    private javax.swing.JPanel jPanel_end_of_fight;
+    private javax.swing.JPanel jPanel_fight;
+    private javax.swing.JPanel jPanel_locations;
+    private javax.swing.JPanel jPanel_records_table;
+    private javax.swing.JPanel jPanel_start;
+    private javax.swing.JPanel jPanel_unavailable_item;
+    private javax.swing.JPanel jPanel_win_in_top;
+    private javax.swing.JPanel jPanel_win_outside_top;
+    private javax.swing.JProgressBar jProgressBar_enemy;
+    private javax.swing.JProgressBar jProgressBar_self;
+    private javax.swing.JRadioButton jRadioButton_big_potion;
+    private javax.swing.JRadioButton jRadioButton_resurrection_cross;
+    private javax.swing.JRadioButton jRadioButton_small_potion;
+    private javax.swing.JScrollPane jScrollPane_records_table;
+    private javax.swing.JTable jTable_records;
+    private javax.swing.JTextField jTextField_locations;
+    private javax.swing.JTextField jTextField_win_in_top;
+    // End of variables declaration//GEN-END:variables
+}
