+++ src/main/java/MortalCombat/Game/Game.java
@@ -1,0 +1,4 @@
+package MortalCombat.Game;
+
+public class Game implements IGame {
+}
