+++ src/main/java/MortalCombat/Game/IGame.java
@@ -1,0 +1,4 @@
+package MortalCombat.Game;
+
+public interface IGame {
+}
