+++ src/main/java/MortalCombat/Handlers/ImageResizer.java
@@ -1,0 +1,15 @@
+package MortalCombat.Handlers;
+
+import java.awt.*;
+import java.awt.image.BufferedImage;
+
+public class ImageResizer {
+    public static BufferedImage resizeImage(BufferedImage originalImage, int targetWidth, int targetHeight) {
+        Image resultingImage = originalImage.getScaledInstance(targetWidth, targetHeight, Image.SCALE_SMOOTH);
+        BufferedImage outputImage = new BufferedImage(targetWidth, targetHeight, BufferedImage.TYPE_INT_ARGB);
+        Graphics2D g2d = outputImage.createGraphics();
+        g2d.drawImage(resultingImage, 0, 0, null);
+        g2d.dispose();
+        return outputImage;
+    }
+}
