+++ src/main/java/MortalCombat/Main.java
@@ -1,0 +1,9 @@
+package MortalCombat;
+
+import MortalCombat.GUI.MainWindow;
+
+public class Main {
+    public static void main(String[] args) {
+        MainWindow window = new MainWindow();
+    }
+}
