+++ src/main/java/MortalCombat/MortalKombatBVersion.java
@@ -1,0 +1,25 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Main.java to edit this template
+ */
+package MortalCombat;
+
+import MortalCombat.GUI.JFrames;
+
+/**
+ *
+ * @author Мария
+ */
+public class MortalKombatBVersion {
+
+    /**
+     * @param args the command line arguments
+     */
+    public static void main(String[] args) {
+        
+        JFrames app = new JFrames();
+        app.setLocationRelativeTo(null);
+        app.setVisible(true);
+    }
+    
+}
